pyproject.toml
[project]
name = "bf3-ci-tests"
version = "1.0.0"
description = "BF3 DPU CI Test Framework"
requires-python = ">=3.10"

dependencies = [
    "pytest>=7.0",
    "pytest-html>=3.0",
    "pytest-timeout>=2.0",
    "pytest-ordering>=0.6",
    "pytest-xdist>=3.0",
    "paramiko>=3.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "urllib3>=2.0",
]

[tool.setuptools.packages.find]
include = ["bf3_ci*"]

[project.entry-points."pytest11"]
bf3_report = "bf3_ci.plugins.pytest_bf3_report"
bf3_recovery = "bf3_ci.plugins.pytest_bf3_recovery"
